Bug 611 - Update message serialization - multi protocol attributes serialization for update message

+++ bgp/parser-impl/src/main/java/org/opendaylight/protocol/bgp/parser/impl/message/update/LocalPreferenceAttributeParser.java
@@ -34,7 +34,7 @@ public final class LocalPreferenceAttributeParser implements AttributeParser,Att
         byteAggregator.writeByte(UnsignedBytes.checkedCast(TYPE));
         byteAggregator.writeByte(UnsignedBytes.checkedCast(LOCAL_PREFS_LENGTH));
 
-        byteAggregator.writeShort(localPref.getPref().shortValue());
+        byteAggregator.writeInt(localPref.getPref().shortValue());
     }
 
 }
